Move ai constants to chat server
